Moved PVNameTree to widgets package
